refactor: separate out leave resources to a separate data file
