+++ frontend/src/utils/dayjs.js
@@ -4,11 +4,13 @@ import localizedFormat from "dayjs/plugin/localizedFormat"
 import relativeTime from "dayjs/esm/plugin/relativeTime"
 import isToday from "dayjs/plugin/isToday"
 import isYesterday from "dayjs/plugin/isYesterday"
+import isBetween from "dayjs/plugin/isBetween"
 
 dayjs.extend(updateLocale)
 dayjs.extend(localizedFormat)
 dayjs.extend(relativeTime)
 dayjs.extend(isToday)
 dayjs.extend(isYesterday)
+dayjs.extend(isBetween)
 
 export default dayjs
